// src/types/neetCodeTopic.type.ts
export type NeetCodeTopic =
  | 'arrays-hashing'
  | 'two-pointers'
  | 'sliding-window'
  | 'stack'
  | 'binary-search'
  | 'linked-list'
  | 'trees'
  | 'tries'
  | 'heap-priority-queue'
  | 'backtracking'
  | 'graphs'
  | 'advanced-graphs'
  | '1d-dynamic-programming'
  | '2d-dynamic-programming'
  | 'greedy'
  | 'intervals'
  | 'math-geometry'
  | 'bit-manipulation'
  | 'javascript'
  | 'miscellaneous'